requirements.txt
Django==3.2.9
django-appconf==1.0.5
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-select2==8.1.1
flake8==7.0.0
flake8-html==0.4.3
pillow==10.3.0
virtualenv==20.25.1